Move openai stuff out of conftest
