+++ tests/unit/datarobot_drum/drum/language_predictors/python_predictor/test_python_predictor.py
@@ -16,7 +16,7 @@ from datarobot_drum.drum.enum import TargetType
 from datarobot_drum.drum.language_predictors.python_predictor.python_predictor import (
     PythonPredictor,
 )
-from tests.unit.datarobot_drum.drum.conftest import create_completion, create_completion_chunks
+from tests.unit.datarobot_drum.drum.chat_utils import create_completion, create_completion_chunks
 
 
 @pytest.fixture
